+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/SubjectCredentialStore.kt
@@ -10,9 +10,21 @@ import at.asitplus.openid.OAuth2AuthorizationServerMetadata
 import at.asitplus.openid.SupportedCredentialFormat
 import at.asitplus.signum.indispensable.cosef.io.coseCompliantSerializer
 import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
-import at.asitplus.wallet.lib.data.*
+import at.asitplus.wallet.lib.data.AttributeIndex
 import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.*
+import at.asitplus.wallet.lib.data.CredentialScheme
+import at.asitplus.wallet.lib.data.IsoMdocCredentialScheme
+import at.asitplus.wallet.lib.data.IsoMdocFallbackCredentialScheme
+import at.asitplus.wallet.lib.data.SdJwtCredentialScheme
+import at.asitplus.wallet.lib.data.SdJwtFallbackCredentialScheme
+import at.asitplus.wallet.lib.data.SelectiveDisclosureItem
+import at.asitplus.wallet.lib.data.UnknownCredentialScheme
 import at.asitplus.wallet.lib.data.VcDataModelConstants.VERIFIABLE_CREDENTIAL
+import at.asitplus.wallet.lib.data.VcFallbackCredentialScheme
+import at.asitplus.wallet.lib.data.VcJwtCredentialScheme
+import at.asitplus.wallet.lib.data.VerifiableCredential
+import at.asitplus.wallet.lib.data.VerifiableCredentialJws
+import at.asitplus.wallet.lib.data.VerifiableCredentialSdJwt
 import io.ktor.utils.io.core.toByteArray
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -75,7 +87,7 @@ interface SubjectCredentialStore {
     @Serializable
     sealed interface StoreEntry {
         @Deprecated("Use scheme instead")
-        val schemaUri: String
+        val schemaUri: String?
 
         @Deprecated("Use resolveScheme() instead to support fetching remote definitions")
         val scheme: CredentialScheme
@@ -95,7 +107,7 @@ interface SubjectCredentialStore {
             val vc: VerifiableCredentialJws,
             @Deprecated("Use scheme instead")
             @SerialName("schema-uri")
-            override val schemaUri: String,
+            override val schemaUri: String? = null,
             @SerialName("credential-renewal-info")
             override val renewalInfo: CredentialRenewalInfo? = null,
             /** See [VcJwtCredentialScheme.vcType] */
@@ -133,7 +145,7 @@ interface SubjectCredentialStore {
             val disclosures: Map<String, SelectiveDisclosureItem?>,
             @Deprecated("Use scheme instead")
             @SerialName("schema-uri")
-            override val schemaUri: String,
+            override val schemaUri: String? = null,
             @SerialName("credential-renewal-info")
             override val renewalInfo: CredentialRenewalInfo? = null,
             /** See [SdJwtCredentialScheme.sdJwtType] */
@@ -163,7 +175,7 @@ interface SubjectCredentialStore {
             val issuerSigned: IssuerSigned,
             @Deprecated("Use scheme instead")
             @SerialName("schema-uri")
-            override val schemaUri: String,
+            override val schemaUri: String? = null,
             @SerialName("credential-renewal-info")
             override val renewalInfo: CredentialRenewalInfo? = null,
             /** See [IsoMdocCredentialScheme.isoDocType] */