+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/data/StaticCredentialMetadataRegistryTest.kt
@@ -30,12 +30,10 @@ val StaticCredentialMetadataRegistryTest by matrixSuite {
         )
 
         AttributeIndex.resolveSdJwtAttributeType(vct.string).shouldNotBeNull().apply {
-            schemaUri shouldBe loadedFrom
             sdJwtType shouldBe vct.string
         }
 
         AttributeIndex.resolveIdentifier(vct.string, SD_JWT).apply {
-            schemaUri shouldBe loadedFrom
             sdJwtType shouldBe vct.string
         }
     }
@@ -76,12 +74,10 @@ val StaticCredentialMetadataRegistryTest by matrixSuite {
         )
 
         AttributeIndex.resolveAttributeType(vcType).shouldNotBeNull().apply {
-            schemaUri shouldBe loadedFrom
             this.vcType shouldBe vcType
         }
 
         AttributeIndex.resolveIdentifierPlainJwt(listOf("VerifiableCredential", vcType)).apply {
-            schemaUri shouldBe loadedFrom
             this.vcType shouldBe vcType
         }
     }
@@ -110,13 +106,11 @@ val StaticCredentialMetadataRegistryTest by matrixSuite {
         )
 
         AttributeIndex.resolveIsoDoctype(docType).shouldNotBeNull().apply {
-            schemaUri shouldBe loadedFrom
             isoDocType shouldBe docType
             isoNamespace shouldBe namespace
         }
 
         AttributeIndex.resolveIdentifier(docType, ISO_MDOC).apply {
-            schemaUri shouldBe loadedFrom
             isoDocType shouldBe docType
             isoNamespace shouldBe namespace
         }